requirements.txt
PyGithub==1.53
pandas==1.1.4
seaborn==0.11.0
streamlit==0.71.0
tk
matplotlib